Add support for custom sync flags
fbset discards unknown sync bits in the fb_var_screeninfo structure.

This works ok on the majority of hardware, but it does not work if the
particular framebuffer driver uses some driver-specific proprietary
bits in the sync field. This situation was seen in the wild with a
Freescale framebuffer driver. Using fbset on it for any other purpose
than just to display current framebuffer configuration always resulted
in pixel glitches on the display. This was caused by fbset dropping
the proprietary sync bits.

This patch resolves the issue by extending fbset to:

* losslessly round-trip the sync value along
  struct fb_var_screeninfo → struct VideoMode → struct fb_var_screeninfo

* allow for explicit switching of the custom bits via newly-added -sync
  argument.

Acked-by: Geert Uytterhoeven <geert@linux-m68k.org>
Signed-off-by: Guillem Jover <guillem@debian.org>